move: ChatList 파일 이동 (#17)
